ptr: Expose pointer types for codegen
