+++ ptr/doc.go
@@ -1,5 +1,5 @@
 // Package ptr provides utilities for converting scalar literal type values to and from pointers inline.
 package ptr
 
-//go:generate go run generate.go
+//go:generate go run -tags codegen generate.go
 //go:generate gofmt -w -s .
